program05-download/program05/SchellingSimulation.java
import student.media.*;
import java.awt.Color;
import java.util.Random;

//-------------------------------------------------------------------------
/**
 *  This class is called "SchellingSimulation" 
 *  that extends the "Picture" class. 
 *  It contains several fields, including 
 *  "satisfactionThreshold" and "redLine".
 *  This class is created
 *  Follow it with additional details about its purpose, what abstraction
 *  it represents, and how to use it.
 *
 *  @author thulasiramanmt
 *  @version 2023.10.03
 */
public class SchellingSimulation extends Picture
{
    //~ Fields ................................................................
    private double satisfactionThreshold;
    private int redLine;
    //~ Constructor ...........................................................

    // ----------------------------------------------------------
    /**
     * Initializes a newly created SchellingSimulation object.
     * 
     * @param width  The width of the simulation grid.
     * @param height The height of the simulation grid.

     */
    public SchellingSimulation(int width, int height)
    {
        super(width, height);
        /*# Do any work to initialize your class here. */
        satisfactionThreshold = 0.3;
        redLine = 0;
        
        //Initialize the entire image to be white
    }
    /**
     * A getter method that 
     * returns the satisfaction threshold (a double).
     * 
     * @return The satisfaction threshold value.
     */
    public double getSatisfactionThreshold()
    {
        return satisfactionThreshold;
    }
    /**
     * A setter method that takes a double parameter 
     * and changes the satisfaction 
     * threshold to the specified value.
     * 
     * @param threshold The new statusfaction threshold value
     * to be set.
     */
    public void setSatisfactionThreshold(double threshold) 
    {
        this.satisfactionThreshold = threshold;
    }
    /**
     * A getter method that 
     * returns the redline value (an int).
     * 
     * @return The redline value as an integer.
     */
    public int getRedLine() 
    {
        return redLine;
    }
    /**
     * A setter method that
     * takes an integer parameter and changes 
     * the redline to the specified value.
     *     
     * @param redLine The new redline value to be set. 
     */
    public void setRedLine(int redLine) 
    {
        redLine = redLine;
    }
    /**
     * Takes two double parameters between 0.0 - 1.0 that 
     * represent percentages. This method will "paint" 
     * the image with a randomized collection of blue 
     * and orange pixels. The first parameter represents 
     * the probabilty of any pixel in the image being set 
     * to blue (representing an X agent). 
     * The second represents the probability of any pixel 
     * in the image (below the redline) being set to orange 
     * (representing an O agent). Other pixels will be left
     * alone (empty, represented by  white ).
     * 
     * @param blueProbability The probability 
     * (between 0.0 and 1.0) of any pixel in the 
     * image being set to blue
     * (representing an X agent).
     * 
     * @param orangeProbability The probability 
     * (between 0.0 and 1.0) of any pixel in 
     * the image (below the redline)
     * being set to orange (representing an O agent).

     * 
     */
    public void populate(double blueProbability, double orangeProbability) 
    {
    }
    /**
     * Takes two Pixel objects and returns a boolean 
     * value indicating whether the two pixels have 
     * the same color.
     * 
     * @param pixel1 The first Pixel object for comparison.
     * @param pixel2 The second Pixel object for comparison.
     * 
     * @return true if the two pixels have the same color. 
     */
    public boolean areSameColor(Pixel pixel1, Pixel pixel2) 
    {
        // stubb
        return true;
    }
    /**
     * Takes one Pixel object and returns true 
     * if its color is Color.WHITE, representing an 
     * empty location.
     * 
     * @param pixel The Pixel object to be checked for emptiness.
     * 
     * @return true if the pixel's color is Color.WHITE (empty).
     */
    public boolean isEmpty(Pixel pixel) 
    {
        // stubb
        return true;
    }
    /**
     * Takes one Pixel object and a Color value, 
     * and returns a boolean result indicating 
     * whether an agent of the specified Color 
     * would be satisfied at the given Pixel location. 
     * It should return true if there are no 
     * neighboring agents, or if the proportion 
     * of neighboring agents the same color as 
     * the parameter meets or exceeds the 
     * satisfaction threshold. Remember 
     * not to count empty neighboring pixels, 
     * only neighboring agents.
     * 
     * @param pixel The Pixel object 
     * representing the location.
     * 
     * @param agentColor The Color value representing 
     * the agent's color.
     * 
     * @return true if there are no neighboring agents or 
     * if the proportion of neighboring agents of the same color
     * as the parameter meets or exceeds the satisfaction 
     * threshold.
     */
    public boolean isSatisfied(Pixel pixel, Color agentColor) 
    {
        // Implementation code here
        return true;
    }
    /**
     * Takes one Pixel object as a parameter 
     * and returns a boolean result. 
     * This method tries to move the corresponding agent 
     * to a new pixel location. This involves picking 
     * a new random x/y location in the 
     * image (below the redline, if the current 
     * pixel is occupied by an orange agent). If 
     * the selected location is empty, and if the 
     * current agent would be satisfied at the new 
     * location, then "move" the agent by setting the 
     * new location's color to the current pixel's 
     * color and setting the current pixel's color to Color.
     * WHITE. Otherwise, leave the current agent 
     * in place, if it cannot be moved to the 
     * selected destination. Return true if the 
     * agent is moved, or false if the attempt 
     * to move fails.
     * 
     * @param pixel The Pixel object 
     * representing the current agent's location.
     * 
     * @return true if the agent is successfully moved
     */
    public boolean maybeRelocate(Pixel pixel) 
    {
        // stubb
        return true;
    }
    /**
     * This method uses a loop over all pixels, 
     * checking each pixel in turn. 
     * If the pixel is empty, leave it alone. 
     * If it is occupied, check to see if the 
     * agent there is satisfied--if so, 
     * leave the agent alone. Otherwise, attempt to 
     * relocate it to a new position 
     * (which could succeed or fail, as 
     * described under maybeRelocate(). 
     * The method returns the total number of 
     * successful moves after processing all pixels.
     * 
     * @return The total number of successful
     * agent moves after processing all pixels.
     */
    public int cycleAgents() 
    {
        // stubb
        return 0;
    }

    //~ Methods ...............................................................


}
